Bug 1558345 - move provider-related endpoints to provider source file
